requirements.txt
pillow>=10.3.0
pyinstaller>=6.9.0